+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/shortTermWeatherForecast/writer/entity/WeatherRiskRepository.java
@@ -1,0 +1,7 @@
+package com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.writer.entity;
+
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface WeatherRiskRepository extends JpaRepository<WeatherRisk, Long> {
+}
