--- batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/writer/entity/ShortTermWeatherForecastId.java
@@ -1,16 +1,0 @@
-package com.ImSnacks.NyeoreumnagiBatch.writer.entity;
-
-import lombok.*;
-
-import java.io.Serializable;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode
-public class ShortTermWeatherForecastId implements Serializable {
-    private int nx;
-    private int ny;
-    private int fcstTime;
-}
