--- batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/writer/entity/WeatherRiskRepository.java
@@ -1,7 +1,0 @@
-package com.ImSnacks.NyeoreumnagiBatch.writer.entity;
-
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface WeatherRiskRepository extends JpaRepository<WeatherRisk, Long> {
-}
