--- batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/writer/repository/ShortTermWeatherForecastRepository.java
@@ -1,8 +1,0 @@
-package com.ImSnacks.NyeoreumnagiBatch.writer.repository;
-
-import com.ImSnacks.NyeoreumnagiBatch.writer.entity.ShortTermWeatherForecastId;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ShortTermWeatherForecastRepository extends JpaRepository<com.ImSnacks.NyeoreumnagiBatch.writer.entity.ShortTermWeatherForecast, ShortTermWeatherForecastId> {
-
-}
